requirements.txt
openai
PyPDF2
python-dotenv
flask
google-generativeai
supabase